# XWIKI-11112: Adding a rendering cache to the Icon Theme Application