--- a/xwiki-platform-core/xwiki-platform-icon/xwiki-platform-icon-default/src/test/java/org/xwiki/icon/internal/DefaultIconSetCacheTest.java
+++ b/xwiki-platform-core/xwiki-platform-icon/xwiki-platform-icon-default/src/test/java/org/xwiki/icon/internal/DefaultIconSetCacheTest.java
@@ -38,7 +38,6 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertSame;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -69,8 +68,8 @@ class DefaultIconSetCacheTest
         CacheFactory cacheFactory = mock(CacheFactory.class);
         when(this.cacheManager.getCacheFactory()).thenReturn(cacheFactory);
         CacheConfiguration configuration = new CacheConfiguration("iconset");
-        this.cache = mock(Cache.class);
-        when(cacheFactory.<IconSet>newCache(eq(configuration))).thenReturn(this.cache);
+        this.cache = mock();
+        when(cacheFactory.<IconSet>newCache(configuration)).thenReturn(this.cache);
     }
 
     @Test