# Jira URL

<!-- Add the link to the corresponding JIRA issue referenced in a commit message. Unless this is a [Misc] commit,
see https://dev.xwiki.org/xwiki/bin/view/Community/DevelopmentPractices#HRule:Don27tcreateunnecessaryissues
-->

https://jira.xwiki.org/browse/XWIKI-11112

# Changes

## Description

<!-- Describe the main changes brought in this PR. -->

* Introduce the internal concept of an `IconTemplateRenderer`.
* Add an internal IconTemplateRendererManager that provides icon template renderers and caches them.
* Cache parsed Velocity code.
* Add special cases for the most commonly used Velocity scripts to avoid expensive Velocity execution for those cases.
* Adapt and add tests.
* Decrease the expected code coverage to 0.96 as 0.97 is just unrealistic.
* Don't ask for non-existing components in the extended URL reference serializer. Instead, check if the component exists, first. This speeds up getting the WebJar URL for Font Awesome by about a factor of 5.

## Clarifications

<!-- Provide extra hints to make it easier to understand the PR. Those could be:
* Explanation of choices made in this PR
* Anchor towards extra resources needed to understand the context of this PR (e.g., a forum proposal).
* Links to other issues this issue depends on
-->

* The last item/second commit could also be a separate Jira or a [Misc] commit, let me know if you want that more formal approach.
* I chose the approach of just caching the template, not caching the rendered output to avoid having to deal with cache invalidations. Further, I've added special cases to avoid Velocity rendering completely for Font Awesome.
* Some numbers of loading all icons for the icon picker (see also https://jira.xwiki.org/browse/XWIKI-22540):

| Icon theme | Before | After |
|------------------|-----------|---------|
| Font Awesome | 283ms | 20ms |
| Silk (first)| 10s | 8.67s |
| Silk (cached) | 770ms | 330ms |

We could decrease the cached running time of Silk further (might be another factor 2) by adding another special case to detect the calls `$xwiki.getSkinFile("icons/silk/${icon}.png")` but I'm not sure if it is worth the extra complexity.

The fraction of icon rendering of page loads went down from approximately 10% to 1% with Font Awesome. The remaining part is all about generating the WebJar URL.

# Screenshots & Video

<!-- If this PR introduces any UI change, it's recommended to highlight it with before/after screenshots 
or even a screen recording for complex interactions. 
-->

No UI changes.

# Executed Tests

<!-- Especially important for regression fixes. 
Indicate how changes were tested (e.g., what maven commands were run to validate them).
-->

Built the changed modules with quality profile. Performed manual tests (see also above).

# Expected merging strategy

* Prefers squash: Yes <!-- No — Explain why. -->
* Backport on branches:
  * Not sure, might be nice for performance, but the code is not completely trivial.